refactor: Move normalize_string_series to preprocessing_utils
