+++ jobs/ml_jobs/artist_linkage/extract_from_wikidata.py
@@ -6,7 +6,7 @@ import requests
 import typer
 from loguru import logger
 
-from match_artists_on_wikidata import normalize_string_series
+from utils.preprocessing_utils import normalize_string_series
 
 QLEVER_ENDPOINT = "https://qlever.cs.uni-freiburg.de/api/wikidata"
 QLEVER_HEADERS = {"Accept": "text/csv", "Content-Type": "application/sparql-query"}
